* generic/regc_locale.c (cclass): 	* generic/tclExecute.c (TclExecuteByteCode): 	* generic/tclIOCmd.c (Tcl_ExecObjCmd): 	* generic/tclListObj.c (NewListIntRep): 	* generic/tclObj.c (Tcl_GetLongFromObj, Tcl_GetWideIntFromObj, 	                    FreeBignum, Tcl_SetBignumObj): 	* generic/tclParseExpr.c (Tcl_ParseExpr): 	* generic/tclStrToD.c (TclParseNumber): 	* generic/tclStringObj.c (TclAppendFormattedObjs): 	* unix/tclUnixPipe.c (TclpCreateProcess): fix signed-with-unsigned 	comparison and other warnings from gcc4 -Wextra.

+++ generic/tclListObj.c
@@ -90,7 +90,7 @@ NewListIntRep(
      * requires API changes to fix.
      */
 
-    if (objc > INT_MAX/sizeof(Tcl_Obj *)) {
+    if ((size_t)objc > INT_MAX/sizeof(Tcl_Obj *)) {
 	return NULL;
     }
