+++ generic/tclStringObj.c
@@ -2145,7 +2145,7 @@ TclAppendFormattedObjs(
 		    int digitOffset;
 
 		    if (useBig) {
-			if (shift<CHAR_BIT*sizeof(Tcl_WideUInt)-DIGIT_BIT) {
+			if ((size_t)shift<CHAR_BIT*sizeof(Tcl_WideUInt)-DIGIT_BIT) {
 			    bits |= (((Tcl_WideUInt)big.dp[index++]) << shift);
 			    shift += DIGIT_BIT;
 			}
